Added the project to the repo
This is a technical assesment for a company I was applying to. The idea was to read a maze from the text file, then instantiate the maze into the world. Once that was done to find a solution to the maze and animate a character who would walk the path to the end of the maze.